refactor(BettingHand): PlayerStatus 이름 변경
